python: promote JobInfo class to bindings API
Problem: The JobInfo class in flux-jobs is useful for abstracting the
raw job data returned from the job-info list APIs into a more usable
interface, but it is not available to Python bindings users.

Promote the JobInfo class to flux.job, so it may be used by other
Flux components and users of the Python API.

Fixes #3140
